move `compare_cost` related functions from utils.py to interface/utils.py
