+++ mitiq/__init__.py
@@ -13,7 +13,8 @@ from mitiq.typing import (
 )
 
 # Utils
-from mitiq.utils import compare_cost, qem_methods
+from mitiq.interface import compare_cost
+from mitiq.utils import qem_methods
 
 # Executors and observables.
 from mitiq.executor import Executor
